+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     `requests.Session` instance is desired. `session` also makes it possible to
     write tests for the new kwarg `use_cached`.
   - Added the helper functions `cached_data_info` and `remove_cached_data`.
+* `Reader` has the new `to_strs` method that yields CHAT data strings.
+* `Reader` has the new `to_chat` method that create CHAT data files in a local directory.
 
 ### Changed
 ### Deprecated